api/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
alembic
pydantic-settings
python-multipart
